Global System Hook to toggle visibility
Adds global system hook for CTRL+W to toggle visibility on visible
overlay windows.  For Example: if Mini Parse Overlay is visible, CTRL+W
hides it, pressed again, it shows it.  If Spell Overlay is visible,
same, if both are visible they are both hidden and displayed.

Since its registered as a global hotkey with windows, the shortcut will
work regardless of which window has focus.